acp {amap}	R Documentation

Principal component analysis

Description

Principal component analysis

Usage

acp(x,center=TRUE,reduce=TRUE,wI=rep(1,nrow(x)),wV=rep(1,ncol(x)))
pca(x,center=TRUE,reduce=TRUE,wI=rep(1,nrow(x)),wV=rep(1,ncol(x)))
## S3 method for class 'acp'
print(x, ...) 

Arguments

x
Matrix / data frame

center
a logical value indicating whether we center data

reduce
a logical value indicating whether we "reduce" data i.e. divide each column by standard deviation

wI,wV
weigth vector for individuals / variables

...
arguments to be passed to or from other methods.

Details

This function offer a variant of princomp and prcomp functions, with a slightly different graphic representation (see plot.acp).

Value

An object of class acp The object is a list with components:

sdev
the standard deviations of the principal components.

loadings
the matrix of variable loadings (i.e., a matrix whose columns contain the eigenvectors). This is of class "loadings": see loadings for its print method.

scores
if scores = TRUE, the scores of the supplied data on the principal components.

eig
Eigen values

Author(s)

Antoine Lucas

See Also

plot.acp,acpgen, princomp

Examples

data(lubisch)
lubisch <- lubisch[,-c(1,8)]
p <- acp(lubisch)
plot(p)

[Package amap version 0.8-20 Index]