acpgen {amap}	R Documentation

Generalised principal component analysis

Description

Generalised principal component analysis

Usage

acpgen(x,h1,h2,center=TRUE,reduce=TRUE,kernel="gaussien")
K(u,kernel="gaussien")
W(x,h,D=NULL,kernel="gaussien")

Arguments

x
Matrix or data frame

h
Scalar: bandwidth of the Kernel

h1
Scalar: bandwidth of the Kernel for W

h2
Scalar: bandwidth of the Kernel for U

kernel
The kernel used. This must be one of '"gaussien"', '"quartic"', '"triweight"', '"epanechikov"' , '"cosinus"' or '"uniform"'

center
A logical value indicating whether we center data

reduce
A logical value indicating whether we "reduce" data i.e. divide each column by standard deviation

D
A product scalar matrix / une matrice de produit scalaire

u
Vector

Details

acpgen compute generalised pca. i.e. spectral analysis of Un / Wn, and project Xi with 1/Wn on the principal vector sub-spaces.

Xi a column vector of p variables of individu i (input data)

W compute estimation of noise in the variance.

W: see latex doc

with Vn variance estimation;

U compute robust variance. 1/Un = 1/Sn - 1 / (h Vn)

S: see latex doc

with μ_n estimator of the mean.

K compute kernel, i.e.

gaussien:

1/sqrt(2pi) exp(-u^2/2)

quartic:

15/16 (1-u^2)^2 if |u| < 1

triweight:

35/32 (1-u^2)^3 if |u| < 1

epanechikov:

3/4 (1-u^2) if |u| < 1

cosinus:

pi/4 cos (u * pi/2) if |u| < 1

Value

An object of class acp The object is a list with components:

sdev
the standard deviations of the principal components.

loadings
the matrix of variable loadings (i.e., a matrix whose columns contain the eigenvectors). This is of class "loadings": see loadings for its print method.

scores
if scores = TRUE, the scores of the supplied data on the principal components.

eig
Eigen values

Author(s)

Antoine Lucas

References

H. Caussinus, M. Fekri, S. Hakam and A. Ruiz-Gazen, A monitoring display of multivariate outliers Computational Statistics & Data Analysis, Volume 44, Issues 1-2, 28 October 2003, Pages 237-252

Caussinus, H and Ruiz-Gazen, A. (1993): Projection Pursuit and Generalized Principal Component Analyses, in New Directions in Statistical Data Analysis and Robustness (eds. Morgenthaler et al.), pp. 35-46. Birk\"auser Verlag Basel.

Caussinus, H. and Ruiz-Gazen, A. (1995). Metrics for Finding Typical Structures by Means of Principal Component Analysis. In Data Science and its Applications (eds Y. Escoufier and C. Hayashi), pp. 177-192. Tokyo: Academic Press.

Antoine Lucas and Sylvain Jasson, Using amap and ctc Packages for Huge Clustering, R News, 2006, vol 6, issue 5 pages 58-60.

See Also

acp acprob princomp

Examples

data(lubisch)
lubisch <- lubisch[,-c(1,8)]
p <- acpgen(lubisch,h1=1,h2=1/sqrt(2))
plot(p,main='ACP robuste des individus')

# See difference with acp

p <- princomp(lubisch)
class(p)<- "acp"

[Package amap version 0.8-20 Index]