afc {amap}	R Documentation

Correspondance factorial analysis.

Description

Compute an acp on a contingency table tacking into account weight of rows and columns

Usage

afc(x)

Arguments

x
A contingency table, or a result of function burt or matlogic

Author(s)

Antoine Lucas

Examples

## Not run: 
color <- as.factor(c('blue','red','red','blue','red'))
size <- as.factor(c('large','large','small','medium','large'))
x <- data.frame(color,size)

afc.1 <- afc(burt(x))
afc.2 <- afc(matlogic(x))

plotAll(afc.1)
plotAll(afc.2)

## End(Not run)

[Package amap version 0.8-20 Index]