Dist {amap}	R Documentation

Distance Matrix Computation

Description

This function computes and returns the distance matrix computed by using the specified distance measure to compute the distances between the rows of a data matrix.

Usage

Dist(x, method = "euclidean", nbproc = 2, diag = FALSE, upper = FALSE)

Arguments

x
numeric matrix or (data frame) or an object of class "exprSet". Distances between the rows of x will be computed.

method
the distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "pearson", "abspearson", "correlation", "abscorrelation", "spearman" or "kendall". Any unambiguous substring can be given.

nbproc
integer, Number of subprocess for parallelization

diag
logical value indicating whether the diagonal of the distance matrix should be printed by print.dist.

upper
logical value indicating whether the upper triangle of the distance matrix should be printed by print.dist.

Details

Available distance measures are (written for two vectors x and y):

euclidean:

Usual square distance between the two vectors (2 norm).

maximum:

Maximum distance between two components of x and y (supremum norm)

manhattan:

Absolute distance between the two vectors (1 norm).

canberra:

sum(|x_i - y_i| / |x_i + y_i|). Terms with zero numerator and denominator are omitted from the sum and treated as if the values were missing.

binary:

(aka asymmetric binary): The vectors are regarded as binary bits, so non-zero elements are ‘on’ and zero elements are ‘off’. The distance is the proportion of bits in which only one is on amongst those in which at least one is on.

pearson:

Also named "not centered Pearson" 1 - sum(x_i y_i) / sqrt [sum(x_i^2) sum(y_i^2)].

abspearson:

Absolute Pearson 1 - |sum(x_i y_i) / sqrt [sum(x_i^2) sum(y_i^2)] |.

correlation:

Also named "Centered Pearson" 1 - corr(x,y).

abscorrelation:

Absolute correlation 1 - | corr(x,y) | with

corr(x,y) = \frac{∑_i x_i y_i -\frac1n ∑_i x_i ∑_i% y_i}{% frac: 2nd part √{≤ft(∑_i x_i^2 -\frac1n ≤ft( ∑_i x_i \right)^2 % \right)% ≤ft( ∑_i y_i^2 -\frac1n ≤ft( ∑_i y_i \right)^2 % \right)} }.

spearman:

Compute a distance based on rank. sum (d_i^2) where d_i is the difference in rank between x_i and y_i.

Dist(x,method="spearman")[i,j] =

cor.test(x[i,],x[j,],method="spearman")$statistic

kendall:

Compute a distance based on rank. ∑_{i,j} K_{i,j}(x,y) with K_{i,j}(x,y) is 0 if x_i, x_j in same order as y_i,y_j, 1 if not.

Missing values are allowed, and are excluded from all computations involving the rows within which they occur. If some columns are excluded in calculating a Euclidean, Manhattan or Canberra distance, the sum is scaled up proportionally to the number of columns used. If all pairs are excluded when calculating a particular distance, the value is NA.

The functions as.matrix.dist() and as.dist() can be used for conversion between objects of class "dist" and conventional distance matrices and vice versa.

Value

An object of class "dist".

The lower triangle of the distance matrix stored by columns in a vector, say do. If n is the number of observations, i.e., n <- attr(do, "Size"), then for i < j <= n, the dissimilarity between (row) i and j is do[n*(i-1) - i*(i-1)/2 + j-i]. The length of the vector is n*(n-1)/2, i.e., of order n^2.

The object has the following attributes (besides "class" equal to "dist"):

Size
integer, the number of observations in the dataset.

Labels
optionally, contains the labels, if any, of the observations of the dataset.

Diag, Upper
logicals corresponding to the arguments diag and upper above, specifying how the object should be printed.

call
optionally, the call used to create the object.

methods
optionally, the distance method used; resulting form dist(), the (match.arg()ed) method argument.

Note

Multi-thread (parallelisation) is disable on Windows.

References

Mardia, K. V., Kent, J. T. and Bibby, J. M. (1979) Multivariate Analysis. London: Academic Press.

Wikipedia https://en.wikipedia.org/wiki/Kendall_tau_distance

See Also

daisy in the ‘cluster’ package with more possibilities in the case of mixed (contiuous / categorical) variables. dist hcluster.

Examples

x <- matrix(rnorm(100), nrow=5)
Dist(x)
Dist(x, diag = TRUE)
Dist(x, upper = TRUE)

## compute dist with 8 threads
Dist(x,nbproc=8)

Dist(x,method="abscorrelation")
Dist(x,method="kendall")

[Package amap version 0.8-20 Index]