plot {amap}	R Documentation

Graphics for Principal component Analysis

Description

Graphics for Principal component Analysis

Usage

## S3 method for class 'acp'
plot(x,i=1,j=2,text=TRUE,label='Composants',col='darkblue',
main='Individuals PCA',variables=TRUE,individual.label=NULL,...)
## S3 method for class 'acp'
biplot(x,i=1,j=2,label='Composants',col='darkblue',length=0.1,
main='Variables PCA',circle=TRUE,...)
plot2(x,pourcent=FALSE,eigen=TRUE,label='Comp.',col='lightgrey',
main='Scree Graph',ylab='Eigen Values')
plotAll(x)

Arguments

x
Result of acp or princomp

i
X axis

j
Y axis

text
a logical value indicating whether we use text or points for plot

pourcent
a logical value indicating whether we use pourcentage of values

eigen
a logical value indicating whether we use eigen values or standard deviation

label
label for X and Y axis

individual.label
labels naming individuals

col
Color of plot

main
Title of graphic

ylab
Y label

length
length of arrows

variables,circle
a logical value indicating whether we display circle or variables

...
cex, pch, and other options; see points.

Value

Graphics:

plot.acp PCA for lines (individuals)

plot.acp PCA for columns (variables)

plot2 Eigen values diagram (Scree Graph)

plotAll Plot both 3 graphs

Author(s)

Antoine Lucas

See Also

acpgen,acprob, princomp

Examples

data(lubisch)
lubisch <- lubisch[,-c(1,8)]
p <- acp(lubisch)
plotAll(p)

[Package amap version 0.8-20 Index]